Added database methods : updateSet(String fileName) getSetSize(String fileName) getAnswers(String setName, int cardIndex) getanswer(String setName, int cardIndex) getQuestion(String setName, int cardIndex) addCreateSetListItem(CreateSetListItem c) ClearCreateSetListItem() getCardAtIndex(String setName, int cardIndex)
These methods are used in FXController.java to get information from the database.

Changes in CreateSetListItem.java :
added methods getTermMCS() and getDefinitionsMCS(). These methods are called
when saving created multiple choice sets. This because the terms and
definition objects are different objects depending on if it's a flash /
spelling set or a multiple choice set.

Changes in FXController.java :

Redefenition changes and minor changes -
118 : setUpFlashcardPlay() - added getQuestion(curSetName,0)
132 : nextFlashCard() - added getQuestion(curSetName,atQuestionInd)
148 : triggerSpinCardAnimation() - Initiate animation with current text
instead of always the defenition
161 : previousFlashCard() - made it so confirmbutton switches its text to
"next" when clicking the previous button.
164 : previousFlashCard() - added getQuestion(curSetName, atQuestionInd)
182 : turnCard() - added getAnswer(curSetName,atQuestionInd)
185 : turnCard() - added getQuestion(curSetName, atQuestionInd)
204 : addCreateSetItem() - added addCreateSetListItem(item)
237 : openCreateSet() added getCreateSetListItems()

saveCreatedSet() Added if statement that makes sure a set with the same
name has not already been created. Implemented two different ways of saving
depending on the setType with if statements. Error messages for each
error added.

setUpMultiplePlay() - Added userAnswers.clear and
resultFlowPane.getChildren().clear() to top of method.

setAnswers() - changed to using the Collections.shuffle method
instead of implementing a shuffle using Random.

presentResult : Fixed the display of the result percentage.

Comment : When pressing the text on the "next" button when playing
multiple choice sets, the answer doesn't get added to the userAnswers
list and is therefore not registered in the result. Quick fix
- remove the on click on the text.