Changes in FXController.java : Added setUpPlay() method that is used in all setup methods. Added nextCard() methods that is used in all nextcard methods These methods reduce code duplication.
Changes in learno.fxml :
Bugfix : Fixed so that clicking the text "next" when playing a
multiple choice set registers the answer.

Changes in mySetsListItem.fxml :
Changed the remove set picture from a bin image to the cross. Did this
because the previous bin image was missing from the images folder.